Rename acl-controller -> controller
